Control: add hysteresis post-filtering to Root CV
This fixes the noise issue with noisy root CV signals (4fdafbb). The noise was mostly coming from the CV (the pot is already heavily filtered), since it was only smoothed by a mild Average filter. Adding a hysteresis filter after it seems to solve the issue. The hysteresis "plateau" is arbitrarily set to 1/10th of a semitone; it should be set to the maximum amplitude of the noise coming from Root, setting it larger will decrease the preciseness of Root CV.