Control: make knobs inactive while Shift is pressed
This supersedes (I think/hope) DG's caching solution, and fixes the issue of main pot values moving between the moment Shift is pressed and the moment a PotMove event is received (which puts the pot in its alternate function).

The new logic for dual-function pots is:
- when Shift is pressed, we stop updating/forwarding the value of all pots with a shift-function (disable(), or put into INACTIVE state); notably, its main value does not change.
- if we move one of these pots, it goes into ALT state, and we now update the alt function.
- when Shift is released, all pots return to their main function (call to main()):
    - the ones who have not moved are reactivated (INACTIVE -> MAIN)
    - the one who have moved go into catchup state (ALT -> CATCHUP)

The same reasoning holds for Learn mode.