Store the Pitch pot dual function (alt/main) state in AltParameters
Pitch pot state data is not used to validate the packet read from Flash, in order to be compatible with the layout from earlier firmware versions lacking this data. Instead, the pitch pot state data is validated by the DualFunctionPot ctor.

Default is Main mode, as Catchup mode is stored as an improbable value